# docs/sasist-agent/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Sasist Agent Protocol API
  version: "1.0.0"
  description: |
    HTTPS surface for Sasist Agent (protocol_version = 1).
    Primary realtime channel is WebSocket — see ws-protocol-v1.md.
    Legacy `/api/printing/*` remains as protocol 0 compatibility until Cleanup.
servers:
  - url: /api/agent
    description: Edge Device Management APIs (RC — actual mount)
  - url: /api/printing
    description: Compat printing register/jobs/heartbeat (RC adapter)
  - url: /api/agent/v1
    description: Planned native agent protocol (not mounted in RC)
...
paths:
  /agents/register:
    post:
      tags: [register]
      summary: Register or re-register agent for this machine
      description: |
        Exchange integration API key for agent session token.
        Upserts by (tenant_id, machine_id).
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRegisterRequest"
            example:
              machine_id: "WIN-DESKTOP-A1B2"
              name: "Pakowanie 1"
              agent_version: "1.0.0"
              protocol_version: 1
              warehouse_id: 3
              host_name: "PAK-01"
              lan_ip: "192.168.10.42"
              os_version: "Windows 11 24H2"
              modules:
                - module_id: printing
                  module_version: "1.0.0"
                  capabilities: ["print.pdf", "print.zpl", "print.raw"]
              devices:
                - local_id: "Zebra ZD420"
                  display_name: "Zebra etykiety"
                  device_kind: printer
                  module_id: printing
                  is_default: true
                  capabilities: ["zpl", "usb"]
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRegisterResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /agents/heartbeat:
    post:
      tags: [heartbeat]
      summary: Heartbeat (HTTPS fallback when WS is down)
      security:
        - AgentTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentHeartbeatRequest"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentHeartbeatResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "426":
          description: Protocol version not supported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolRejected"
...
  /agents/{agent_id}/diagnostics:
    get:
      tags: [diagnostics]
      summary: Latest diagnostics report (ERP user session)
      security:
        - UserSessionAuth: []
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Latest report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticsReport"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /agents/devices:
    get:
      tags: [devices]
      summary: List devices known for this agent session
      security:
        - AgentTokenAuth: []
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"

    put:
      tags: [devices]
      summary: Full replace device inventory from agent
      security:
        - AgentTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [devices]
              properties:
                devices:
                  type: array
                  items:
                    $ref: "#/components/schemas/Device"
      responses:
        "204":
          description: Replaced

  # Parallel ERP operator surface (migration alongside /api/printing/*)
  /devices:
    get:
      tags: [devices]
      summary: List edge devices for tenant (operator UI) — parallel to /printing/printers
      description: |
        Mounted in ERP as `/api/agent/devices`. Type-agnostic projection.
        Does not replace agent-authenticated `/agents/devices`.
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
        - name: warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: agent_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/DeviceKind"
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EdgeDevice"
  /device/{device_id}:
    get:
      tags: [devices]
      summary: Get one edge device by id / system_name / legacy printer id
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string }
        - name: tenant_id
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EdgeDevice"
        "404":
          description: Not found
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Integration API key (`spa_` / `sasist_`) with agent scope
    AgentTokenAuth:
      type: http
      scheme: bearer
      description: Agent access token (`spt_` compat / `sat_` preferred)
    AgentRefreshAuth:
      type: http
      scheme: bearer
      description: Agent refresh token
    UserSessionAuth:
      type: http
      scheme: bearer
      description: ERP user session
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ProtocolRejected:
      type: object
      required: [accepted, min_protocol_version, max_protocol_version, message]
      properties:
        accepted:
          type: boolean
          enum: [false]
        min_protocol_version:
          type: integer
        max_protocol_version:
          type: integer
        message:
          type: string
    UpdateChannel:
      type: string
      enum: [stable, beta]
    ModuleInfo:
      type: object
      required: [module_id, module_version]
      properties:
        module_id:
          type: string
        module_version:
          type: string
        state:
          type: string
          enum: [running, degraded, stopped, failed]
        capabilities:
          type: array
          items:
            type: string
    Device:
      type: object
      required: [local_id, display_name, device_kind, module_id]
      description: |
        Legacy wire shape (protocol v1). Prefer EdgeDevice for new ERP surfaces.
        `capabilities` may be flat tokens for compat; structured descriptors live on EdgeDevice.
      properties:
        local_id:
          type: string
          maxLength: 255
        display_name:
          type: string
          maxLength: 120
        device_kind:
          $ref: "#/components/schemas/DeviceKind"
        module_id:
          type: string
        is_active:
          type: boolean
          default: true
        is_default:
          type: boolean
          default: false
        capabilities:
          type: array
          items:
            type: string
          description: Flattened capability tokens (compat)
        health:
          $ref: "#/components/schemas/DeviceHealth"
        metadata:
          type: object
          additionalProperties: true
    EdgeDevice:
      type: object
      required: [id, type, status]
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/DeviceKind"
        manufacturer:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        serial_number:
          type: string
          nullable: true
        driver:
          type: string
          nullable: true
        firmware:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/DeviceOperationalStatus"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/CapabilityDescriptor"
        last_seen:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        agent_id:
          type: integer
          nullable: true
        module_id:
          type: string
          nullable: true
        display_name:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_default:
          type: boolean
        legacy_printer_id:
          type: integer
          nullable: true
    CapabilityDescriptor:
      type: object
      required: [name, version, supported_operations]
      properties:
        name:
          type: string
        version:
          type: string
        supported_operations:
          type: array
          items:
            type: string
        limits:
          type: object
          additionalProperties: true
          nullable: true
    DeviceOperationalStatus:
      type: string
      enum: [online, offline, error, warning, busy, idle, unknown]
...
    DeviceKind:
      type: string
      enum: [printer, scanner, scale, camera, rfid, usb, serial, custom]
    DeviceHealth:
      type: object
      properties:
        online:
          type: boolean
        status:
          type: string
          description: Prefer DeviceOperationalStatus; legacy ok|warning|error|unknown still accepted
          enum: [ok, warning, error, unknown, online, offline, busy, idle]
        paper:
          type: string
          enum: [ok, low, empty, unknown]
        toner:
          type: string
          enum: [ok, low, empty, unknown]
        queue_depth:
          type: integer
          minimum: 0
        message:
          type: string
          maxLength: 500
        checked_at:
          type: string
          format: date-time
    AgentRegisterRequest:
      type: object
      required: [machine_id, name, protocol_version]
      properties:
        machine_id:
          type: string
          maxLength: 128
        name:
          type: string
          maxLength: 120
        agent_version:
          type: string
          maxLength: 32
        protocol_version:
          type: integer
          minimum: 1
        warehouse_id:
          type: integer
          nullable: true
        host_name:
          type: string
        lan_ip:
          type: string
        os_version:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleInfo"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
    AgentRegisterResponse:
      type: object
      required: [agent_id, access_token, token_type, expires_in, protocol_version, machine_id]
      properties:
        agent_id:
          type: integer
        access_token:
          type: string
        refresh_token:
          type: string
          description: Present when rotation is enabled
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime seconds
        protocol_version:
          type: integer
        machine_id:
          type: string
        tenant_id:
          type: integer
          nullable: true
        warehouse_id:
          type: integer
          nullable: true
        server_capabilities:
          $ref: "#/components/schemas/ServerCapabilities"
...
    AgentHeartbeatRequest:
      type: object
      required: [protocol_version]
      properties:
        protocol_version:
          type: integer
        agent_version:
          type: string
        name:
          type: string
        lan_ip:
          type: string
        connection_mode:
          type: string
          enum: [ws, poll, mixed]
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleInfo"
        device_count:
          type: integer
          minimum: 0
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
          description: Optional compact or full snapshot
        last_error:
          type: string
          maxLength: 2000
          nullable: true
        metrics:
          type: object
          additionalProperties: true
    AgentHeartbeatResponse:
      type: object
      required: [agent_id, is_online, server_time]
      properties:
        agent_id:
          type: integer
        is_online:
          type: boolean
        server_time:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        config_version:
          type: string
          nullable: true
        update:
          $ref: "#/components/schemas/UpdateHint"
    ServerCapabilities:
      type: object
      properties:
        ws:
          type: boolean
        jobs_formats:
          type: array
          items:
            type: string
            enum: [pdf, html, zpl, raw]
        log_upload:
          type: boolean
        remote_diagnostics:
          type: boolean
        update_channels:
          type: array
          items:
            $ref: "#/components/schemas/UpdateChannel"
...
    DiagnosticsSeverity:
      type: string
      enum: [INFO, WARNING, ERROR]
    DiagnosticsCheck:
      type: object
      required: [id, title, severity, status]
      properties:
        id:
          type: string
        title:
          type: string
        severity:
          $ref: "#/components/schemas/DiagnosticsSeverity"
        status:
          type: string
          enum: [pass, fail, skip]
        message:
          type: string
        duration_ms:
          type: integer
    DiagnosticsReport:
      type: object
      required: [run_id, started_at, finished_at, overall_status, checks]
      properties:
        run_id:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        overall_status:
          type: string
          enum: [pass, fail]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticsCheck"
        agent_version:
          type: string
        protocol_version:
          type: integer
...
    UpdateHint:
      type: object
      properties:
        available:
          type: boolean
        version:
          type: string
        channel:
          $ref: "#/components/schemas/UpdateChannel"
        mandatory:
          type: boolean
...